パラメータの調整を行った

+++ Assets/Scripts/Unit/Parts/BodyData.cs
@@ -11,9 +11,9 @@ public class BodyData : UnitPartsData
     [Tooltip("機体出力")]
     [SerializeField] int[] _output;
     [Tooltip("旋回限界")]
-    [SerializeField] int[] _turnRange;
+    [SerializeField] float[] _turnRange;
     /// <summary> 機体出力 </summary>
     public int[] Output { get => _output; }
-    /// <summary> 移動力 </summary>
-    public int[] TurnRange { get => _turnRange; }
+    /// <summary> 旋回限界 </summary>
+    public float[] TurnRange { get => _turnRange; }
 }
